requirements.txt
altgraph==0.17.5
macholib==1.16.4
numpy==2.4.0
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyQt5==5.15.11
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3
